requirements.txt
streamlit==1.10.0
pandas==1.4.3
numpy==1.23.0
Pillow==9.2.0
